Move from domain.ports to interfaces
