+++ src/main/java/br/com/fiap/soat/techChallenge/application/config/BeanConfig.java
@@ -1,9 +1,9 @@
 package br.com.fiap.soat.techChallenge.application.config;
 
-import br.com.fiap.soat.techChallenge.domain.ports.inbound.*;
-import br.com.fiap.soat.techChallenge.domain.ports.outbound.ClienteRepositoryPort;
-import br.com.fiap.soat.techChallenge.domain.ports.outbound.ProdutoRepositoryPort;
-import br.com.fiap.soat.techChallenge.domain.ports.outbound.PedidoRepositoryPort;
+import br.com.fiap.soat.techChallenge.interfaces.usecases.*;
+import br.com.fiap.soat.techChallenge.interfaces.gateways.ClienteRepositoryPort;
+import br.com.fiap.soat.techChallenge.interfaces.gateways.ProdutoRepositoryPort;
+import br.com.fiap.soat.techChallenge.interfaces.gateways.PedidoRepositoryPort;
 
 import br.com.fiap.soat.techChallenge.usecases.*;
 import org.springframework.context.annotation.Bean;
