+++ src/main/java/br/com/fiap/soat/techChallenge/interfaces/usecases/CadastrarClienteUseCasePort.java
@@ -1,4 +1,4 @@
-package br.com.fiap.soat.techChallenge.domain.ports.inbound;
+package br.com.fiap.soat.techChallenge.interfaces.usecases;
 
 import br.com.fiap.soat.techChallenge.entities.Cliente;
