+++ src/main/java/br/com/fiap/soat/techChallenge/usecases/CadastrarClienteUseCase.java
@@ -1,8 +1,8 @@
 package br.com.fiap.soat.techChallenge.usecases;
 
 import br.com.fiap.soat.techChallenge.entities.Cliente;
-import br.com.fiap.soat.techChallenge.domain.ports.inbound.CadastrarClienteUseCasePort;
-import br.com.fiap.soat.techChallenge.domain.ports.outbound.ClienteRepositoryPort;
+import br.com.fiap.soat.techChallenge.interfaces.usecases.CadastrarClienteUseCasePort;
+import br.com.fiap.soat.techChallenge.interfaces.gateways.ClienteRepositoryPort;
 
 public class CadastrarClienteUseCase implements CadastrarClienteUseCasePort {
     private final ClienteRepositoryPort clienteRepositoryPort;
