+++ src/main/java/br/com/fiap/soat/techChallenge/usecases/CadastrarProdutoUseCase.java
@@ -1,7 +1,7 @@
 package br.com.fiap.soat.techChallenge.usecases;
 import br.com.fiap.soat.techChallenge.entities.Produto;
-import br.com.fiap.soat.techChallenge.domain.ports.inbound.CadastrarProdutoUseCasePort;
-import br.com.fiap.soat.techChallenge.domain.ports.outbound.ProdutoRepositoryPort;
+import br.com.fiap.soat.techChallenge.interfaces.usecases.CadastrarProdutoUseCasePort;
+import br.com.fiap.soat.techChallenge.interfaces.gateways.ProdutoRepositoryPort;
 
 public class CadastrarProdutoUseCase implements CadastrarProdutoUseCasePort {
     private final ProdutoRepositoryPort produtoRepositoryPort;
