+++ src/main/java/br/com/fiap/soat/techChallenge/usecases/ExcluirProdutoUseCase.java
@@ -1,7 +1,7 @@
 package br.com.fiap.soat.techChallenge.usecases;
 
-import br.com.fiap.soat.techChallenge.domain.ports.inbound.ExcluirProdutoUseCasePort;
-import br.com.fiap.soat.techChallenge.domain.ports.outbound.ProdutoRepositoryPort;
+import br.com.fiap.soat.techChallenge.interfaces.usecases.ExcluirProdutoUseCasePort;
+import br.com.fiap.soat.techChallenge.interfaces.gateways.ProdutoRepositoryPort;
 
 import java.util.UUID;
