+++ src/main/java/br/com/fiap/soat/techChallenge/usecases/ObterPedidosUseCase.java
@@ -1,8 +1,8 @@
 package br.com.fiap.soat.techChallenge.usecases;
 
 import br.com.fiap.soat.techChallenge.entities.Pedido;
-import br.com.fiap.soat.techChallenge.domain.ports.inbound.ObterPedidosUseCasePort;
-import br.com.fiap.soat.techChallenge.domain.ports.outbound.PedidoRepositoryPort;
+import br.com.fiap.soat.techChallenge.interfaces.usecases.ObterPedidosUseCasePort;
+import br.com.fiap.soat.techChallenge.interfaces.gateways.PedidoRepositoryPort;
 
 import java.util.List;
