+++ src/main/java/br/com/fiap/soat/techChallenge/usecases/ObterProdutosPorCategoriaUseCase.java
@@ -1,7 +1,7 @@
 package br.com.fiap.soat.techChallenge.usecases;
 import br.com.fiap.soat.techChallenge.entities.Produto;
-import br.com.fiap.soat.techChallenge.domain.ports.outbound.ProdutoRepositoryPort;
-import br.com.fiap.soat.techChallenge.domain.ports.inbound.ObterProdutosPorCategoriaUseCasePort;
+import br.com.fiap.soat.techChallenge.interfaces.gateways.ProdutoRepositoryPort;
+import br.com.fiap.soat.techChallenge.interfaces.usecases.ObterProdutosPorCategoriaUseCasePort;
 
 import java.util.List;
